+++ packages/core/lib/commands/autocomplete/commands/install/index.js
