+++ packages/core/lib/commands/autocomplete/commands/install/run.js
