+++ packages/core/lib/commands/autocomplete/commands/uninstall/index.js
