+++ packages/core/lib/commands/autocomplete/commands/uninstall/meta.js
@@ -1,0 +1,10 @@
+module.exports = {
+  command: "uninstall",
+  description: "Uninstall CLI tab-completion",
+  builder: {},
+  help: {
+    usage: "truffle autocomplete uninstall",
+    options: [],
+    allowedGlobalOptions: []
+  }
+};
