+++ packages/core/lib/commands/autocomplete/commands/uninstall/run.js
