+++ packages/core/lib/commands/autocomplete/helpers.js
