+++ packages/core/lib/commands/autocomplete/index.js
