+++ packages/core/lib/commands/autocomplete/meta.js
@@ -2,11 +2,11 @@ const install = require("./commands/install");
 const uninstall = require("./commands/uninstall");
 
 module.exports = {
-  command: "completion",
+  command: "autocomplete",
   description: false, // Not intended to be used by clients
   builder: {},
   help: {
-    usage: "truffle completion <command>",
+    usage: "truffle autocomplete <command>",
     options: [],
     allowedGlobalOptions: []
   },
