+++ packages/core/lib/commands/autocomplete/run.js
