--- packages/core/lib/commands/completion/commands/uninstall/meta.js
@@ -1,10 +1,0 @@
-module.exports = {
-  command: "install",
-  description: "Install CLI tab-completion",
-  builder: {},
-  help: {
-    usage: "truffle completion install",
-    options: [],
-    allowedGlobalOptions: []
-  }
-};
